requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
flask-redis==0.4.0
Flask-Sessions==0.1.5
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.3
postmarker==1.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
Werkzeug==3.0.1
zipp==3.17.0
